requirements.txt
Flask==3.0.3
pytest==8.3.3
pytest-mock==3.14.0
pdfminer.six==20250327
easyocr==1.7.2
python-docx==1.1.2
pillow==10.2.0
numpy==1.26.4
pandas==1.5.1
openpyxl==3.1.2
xlrd==2.0.1
scikit-learn==1.2.2
huggingface_hub==0.20.3
sentence-transformers==2.2.2
setuptools==80.3.0
gunicorn==23.0.0


